# brevesapiserver/routers.go
/*
 * Breves API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package brevesapiserver

import (
	"fmt"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
	security "github.com/ZolaraProject/library/security"
)

type Route struct {
	Name                string
	Method              string
	Pattern             string
	HandlerFunc         http.HandlerFunc
	RequiredPermissions []string
}

type Routes []Route

func NewRouter(jwtSecretKey string) *mux.Router {
	router := mux.NewRouter().StrictSlash(true).UseEncodedPath()
	for _, route := range routes {
		var handler http.Handler
		handler = route.HandlerFunc
		handler = Logger(handler, route.Name)
		handler = http.HandlerFunc(security.PermissionCheck(handler.ServeHTTP, route.RequiredPermissions, jwtSecretKey, RedisPool))

		router.
			Methods(route.Method).
			Path(route.Pattern).
			Name(route.Name).
			Handler(handler)
	}

	return router
}

func Index(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	fmt.Fprintf(w, "{\"name\": \"Breves API\",\"version\": \"1.0.0\"}")
	w.WriteHeader(http.StatusOK)
}

func Healthz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusOK)
}

var routes = Routes{
	Route{
		"Index",
		"GET",
		"/api/",
		Index,
		[]string{},
	},

	Route{
		"Healthz",
		"GET",
		"/healthz",
		Healthz,
		[]string{},
	},

	Route{
		"GetUserVideos",
		strings.ToUpper("Get"),
		"/api/video",
		GetUserVideos,
		[]string{ "USER", },
	},

	Route{
		"LikeVideos",
		strings.ToUpper("Put"),
		"/api/video",
		LikeVideos,
		[]string{ "USER", },
	},

	Route{
		"UpdateUserVideo",
		strings.ToUpper("Put"),
		"/api/video/{videoId}",
		UpdateUserVideo,
		[]string{ "USER", },
	},
}