# models/model_user_created_response.go
/*
 * Breves API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type UserCreatedResponse struct {

	Message string `json:"message"`

	CreatedId int64 `json:"createdId,omitempty"`
}