# models/model_user_video_in_list.go
/*
 * Breves API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type UserVideoInList struct {

	Title string `json:"title"`

	Subtitle string `json:"subtitle"`

	Likes int64 `json:"likes"`

	Language string `json:"language"`

	Level string `json:"level"`

	Action string `json:"action"`

	VideoUrl string `json:"videoUrl"`

	VideoId string `json:"videoId"`
}