# models/model_user_video_list.go
/*
 * Breves API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type UserVideoList struct {

	UserVideos []UserVideoInList `json:"userVideos,omitempty"`

	Total int64 `json:"total"`
}